requirements.txt
flask
line-bot-sdk
requests
deep-translator==1.11.4
